fix 파일 구조 / DB연결 통일
