+++ src/javaResources/ApplyServlet.java
@@ -1,11 +1,13 @@
+package javaResources;
+
+import javaResources.ApplyDao;
+
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
 import java.io.IOException;
-import java.sql.Date;
 
 import java.sql.*;
