+++ src/javaResources/DB1.java
@@ -1,4 +1,4 @@
-package com.hanaHR.web;
+package javaResources;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
@@ -13,7 +13,7 @@ public class DB1 {
         Connection connection = null;
 
         try {
-            Class.forName("jdbc:mysql://172.16.20.89:3306/hanahr?useUnicode=true&characterEncoding=utf8");
+            Class.forName("com.mysql.cj.jdbc.Driver");
             connection = DriverManager.getConnection(JDBC_URL, USERNAME, PASSWORD);
         } catch (ClassNotFoundException e) {
             e.printStackTrace();
