+++ src/javaResources/MemberServlet.java
@@ -20,12 +20,7 @@ public class MemberServlet extends HttpServlet {
     @Override
     public void init() throws ServletException {
         super.init();
-        try {
-            Class.forName("com.mysql.cj.jdbc.Driver");
-            connection = DriverManager.getConnection("jdbc:mysql://172.16.20.89:3306/hanahr?useUnicode=true&characterEncoding=utf8", "hanaro", "hanaro6666!");
-        } catch (ClassNotFoundException | SQLException e) {
-            e.printStackTrace();
-        }
+        Connection connection = DB1.getConnection();
     }
 
     @Override
