+++ src/javaResources/StatusServlet.java
@@ -19,13 +19,7 @@ public class StatusServlet extends HttpServlet {
     @Override
     public void init() throws ServletException {
         super.init();
-        try {
-            Class.forName("com.mysql.cj.jdbc.Driver");
-            connection = DriverManager.getConnection("jdbc:mysql://172.16.20.89:3306/hanahr?useUnicode=true&characterEncoding=utf8", "hanaro", "hanaro6666!");
-        } catch (ClassNotFoundException | SQLException e) {
-
-            e.printStackTrace();
-        }
+        Connection connection = DB1.getConnection();
     }
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
