+++ src/javaResources/updateCandidate.java
@@ -60,13 +60,11 @@ public class updateCandidate extends HttpServlet {
         boolean memberInterview2Pass = Boolean.parseBoolean(request.getParameter("memberInterview2Pass"));
 
         // 데이터베이스 연결 및 업데이트 작업 수행
-        Connection connection = null;
+        Connection connection = DB1.getConnection();
         PreparedStatement updateStatement = null;
 
         try {
             // 데이터베이스 연결 설정
-            Class.forName("com.mysql.cj.jdbc.Driver");
-            connection = DriverManager.getConnection("jdbc:mysql://172.16.20.89:3306/hanahr?useUnicode=true&characterEncoding=utf8", "hanaro", "hanaro6666!");
             System.out.println("------------------");
             // 업데이트 쿼리 작성 score테이블
             String updateScoreQuery = "UPDATE score set memberPaperScore = ?, memberPaperPass = ?, memberWrittenScore = ?, memberWrittenPass = ?, memberInterview1Score = ?, memberInterview1Pass = ?, memberInterview2Score = ?, memberInterview2Pass = ? WHERE memberNumber = ?";
