+++ web/candidateEdit.jsp
@@ -209,7 +209,7 @@
             <!-- Begin Page Content -->
             <div class="container-fluid">
 
-                <%@ page import="javaBeans.User" %>
+                <%@ page import="javaBeans.UserDTO" %>
                 <%@ page import="java.util.List" %>
                 <%@ page import="java.util.ArrayList" %>
 
@@ -246,9 +246,9 @@
                                 </tr>
                                 </thead>
                                 <tbody>
-                                <% List<User> searchResults = (List<User>) request.getAttribute("searchResults"); %>
+                                <% List<UserDTO> searchResults = (List<UserDTO>) request.getAttribute("searchResults"); %>
                                 <% if (searchResults != null && !searchResults.isEmpty()) { %>
-                                <% for (User user : searchResults) { %>
+                                <% for (UserDTO user : searchResults) { %>
                                 <tr>
 
 <%--                                    <td><%= user.getMemberNumber() %></td>--%>
@@ -283,7 +283,7 @@
                             </table>
 
                             <% if (searchResults != null && !searchResults.isEmpty()) { %>
-                            <% for (User user : searchResults) { %>
+                            <% for (UserDTO user : searchResults) { %>
 
                             <form action="/updateCandidate" method="post" accept-charset="UTF-8">
                             <table >
