+++ web/candidateEmail.jsp
@@ -1,8 +1,8 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8" %>
 <%@ page import="java.sql.*" %>
-<%@ page import="com.hanaHR.web.DB1" %>
-<%@ page import="com.hanaHR.web.EmailUtils" %>
-<%@ page import="com.hanaHR.web.Rq" %>
+<%@ page import="javaResources.DB1" %>
+<%@ page import="javaResources.EmailUtils" %>
+<%@ page import="javaResources.Rq" %>
 
 <!DOCTYPE html>
 <html lang="en">
